+++ tkclasswiz/__init__.py
@@ -27,7 +27,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
 
-__version__ = "1.4.4"
+__version__ = "1.4.5"
 
 
 from .object_frame import *