+++ tkclasswiz/object_frame/frame_base.py
@@ -1,8 +1,6 @@
-from typing import get_args, get_origin, Iterable, Union, Literal, Any, TYPE_CHECKING, TypeVar, Generic
+from typing import get_args, get_origin, Iterable, Union, Literal, Any, TYPE_CHECKING, TypeVar
 from abc import ABC, abstractmethod
-from inspect import isabstract
 from contextlib import suppress
-from itertools import chain, product
 from functools import cache
 
 from ..convert import *
@@ -180,76 +178,6 @@ class NewObjectFrameBase(ttk.Frame, ABC):
             raise TypeError(f"Could not convert '{value}' to any of accepted types.\nAccepted types: '{types}'")
 
         return value
-
-    @classmethod
-    def convert_types(cls, input_type: type):
-        """
-        Type preprocessing method, that extends the list of types with inherited members (polymorphism)
-        and removes classes that are wrapped by some other class, if the wrapper class also appears in
-        the annotations.
-        """
-        def remove_classes(types: list):
-            r = types.copy()
-            for type_ in types:
-                # It's a wrapper of some class -> remove the wrapped class
-                if hasattr(type_, "__wrapped__"):
-                    if type_.__wrapped__ in r:
-                        r.remove(type_.__wrapped__)
-
-                # Abstract classes are classes that don't allow instantiation -> remove the class
-                if isabstract(type_):
-                    r.remove(type_)
-
-            return tuple({a:0 for a in r})
-
-
-        if isinstance(input_type, str):
-            raise TypeError(
-                f"Provided type '{input_type}' is not a type - it is a string!\n"
-                "Potential subscripted type problem?\n"
-                "Instead of e. g., list['type'], try using typing.List['type']."
-            )
-
-        origin = get_origin(input_type)
-        if issubclass_noexcept(origin, Generic):
-            # Patch for Python versions < 3.10
-            input_type.__name__ = origin.__name__
-
-        # Unpack Union items into a tuple
-        if origin is Union or issubclass_noexcept(origin, (Iterable, Generic)):
-            new_types = []
-            for arg_group in get_args(input_type):
-                new_types.append(remove_classes(list(cls.convert_types(arg_group))))
-
-            if origin is Union:
-                return tuple(chain.from_iterable(new_types))  # Just expand unions
-
-            # Process abstract classes and polymorphism
-            new_origins = []
-            for origin in cls.convert_types(origin):
-                if issubclass_noexcept(origin, Iterable):
-                    new_origins.append(origin[tuple(chain.from_iterable(new_types))])
-                elif issubclass_noexcept(origin, Generic):
-                    for comb in product(*new_types):
-                        new_origins.append(origin[comb])
-                else:
-                    new_origins.append(origin)
-
-            return remove_classes(new_origins)
-
-        if input_type.__module__ == "builtins":
-            # Don't consider built-int types for polymorphism
-            # No removal of abstract classes is needed either as builtins types aren't abstract
-            return (input_type,)
-
-        # Extend subclasses
-        subtypes = []
-        if hasattr(input_type, "__subclasses__"):
-            for st in input_type.__subclasses__():
-                subtypes.extend(cls.convert_types(st))
-
-        # Remove wrapped classes (eg. wrapped by decorator) + ABC classes
-        return remove_classes([input_type, *subtypes])
 
     def init_main_frame(self):
         frame_main = ttk.Frame(self)